Move conversion error definitions into the dangerzone package
The conversion error classes used to live under dangerzone/conversion/
which is the code that runs inside the container. Now that the
container image lives in a separate repository, the "client" side needs
its own copy of the error types and shared constants (DEFAULT_DPI,
INT_BYTES, MAX_PAGES) so it can decode the integer error codes that
cross the sandbox boundary.

This is now in dangerzone/conversion_errors.py with the ConversionException
hierarchy and the constants that the client needs. Isolation provider tests
have been modified to use this new location.
