+++ dangerzone/isolation_provider/base.py
@@ -16,8 +16,8 @@ from typing import IO, Callable, Iterator, Optional
 import fitz
 from colorama import Fore, Style
 
-from ..conversion import errors
-from ..conversion.common import DEFAULT_DPI, INT_BYTES
+from .. import conversion_errors as errors
+from ..conversion_errors import DEFAULT_DPI, INT_BYTES
 from ..document import Document
 from ..util import get_tessdata_dir, replace_control_chars
